Break down original details form store
