+++ src/components/shared/Entity/Actions/SaveButton.tsx
@@ -5,6 +5,7 @@ import LogDialogActions from 'components/shared/Entity/LogDialogActions';
 import {
     DetailsFormStoreNames,
     DraftEditorStoreNames,
+    FormStateStoreNames,
     useZustandStore,
 } from 'context/Zustand';
 import { DraftSpecQuery } from 'hooks/useDraftSpecs';
@@ -12,7 +13,7 @@ import { useRouteStore } from 'hooks/useRouteStore';
 import { FormattedMessage } from 'react-intl';
 import { CustomEvents } from 'services/logrocket';
 import { entityCreateStoreSelectors } from 'stores/Create';
-import { CreateState, FormStatus } from 'stores/MiniCreate';
+import { EntityFormState, FormStatus } from 'stores/FormState';
 
 interface Props {
     closeLogs: Function;
@@ -21,6 +22,7 @@ interface Props {
     logEvent: CustomEvents.CAPTURE_CREATE | CustomEvents.MATERIALIZATION_CREATE;
     draftEditorStoreName: DraftEditorStoreNames;
     detailsFormStoreName: DetailsFormStoreNames;
+    formStateStoreName: FormStateStoreNames;
     materialize?: Function;
 }
 
@@ -31,33 +33,35 @@ function EntitySaveButton({
     draftEditorStoreName,
     materialize,
     logEvent,
+    formStateStoreName,
     detailsFormStoreName,
 }: Props) {
-    const useEntityCreateStore = useRouteStore();
-
-    const showLogs = useZustandStore<
-        CreateState,
-        CreateState['formState']['showLogs']
-    >(detailsFormStoreName, (state) => state.formState.showLogs);
-
-    const logToken = useZustandStore<
-        CreateState,
-        CreateState['formState']['logToken']
-    >(detailsFormStoreName, (state) => state.formState.logToken);
-
-    const formStatus = useZustandStore<
-        CreateState,
-        CreateState['formState']['status']
-    >(detailsFormStoreName, (state) => state.formState.status);
-
-    const messagePrefix = useEntityCreateStore(
-        entityCreateStoreSelectors.messagePrefix
-    );
-
+    // Draft Editor Store
     const draftId = useZustandStore<
         EditorStoreState<DraftSpecQuery>,
         EditorStoreState<DraftSpecQuery>['id']
     >(draftEditorStoreName, (state) => state.id);
+
+    // Form State Store
+    const showLogs = useZustandStore<
+        EntityFormState,
+        EntityFormState['formState']['showLogs']
+    >(formStateStoreName, (state) => state.formState.showLogs);
+
+    const logToken = useZustandStore<
+        EntityFormState,
+        EntityFormState['formState']['logToken']
+    >(formStateStoreName, (state) => state.formState.logToken);
+
+    const formStatus = useZustandStore<
+        EntityFormState,
+        EntityFormState['formState']['status']
+    >(formStateStoreName, (state) => state.formState.status);
+
+    const useEntityCreateStore = useRouteStore();
+    const messagePrefix = useEntityCreateStore(
+        entityCreateStoreSelectors.messagePrefix
+    );
 
     return (
         <>
@@ -77,7 +81,7 @@ function EntitySaveButton({
                 actionComponent={
                     <LogDialogActions
                         close={closeLogs}
-                        detailsFormStoreName={detailsFormStoreName}
+                        formStateStoreName={formStateStoreName}
                         materialize={
                             materialize
                                 ? {
@@ -94,6 +98,7 @@ function EntitySaveButton({
                 onFailure={callFailed}
                 logEvent={logEvent}
                 draftEditorStoreName={draftEditorStoreName}
+                formStateStoreName={formStateStoreName}
                 detailsFormStoreName={detailsFormStoreName}
             />
         </>
