+++ src/components/shared/Entity/Actions/TestButton.tsx
@@ -5,6 +5,7 @@ import LogDialogActions from 'components/shared/Entity/LogDialogActions';
 import {
     DetailsFormStoreNames,
     DraftEditorStoreNames,
+    FormStateStoreNames,
     useZustandStore,
 } from 'context/Zustand';
 import { DraftSpecQuery } from 'hooks/useDraftSpecs';
@@ -12,7 +13,7 @@ import { useRouteStore } from 'hooks/useRouteStore';
 import { FormattedMessage } from 'react-intl';
 import { CustomEvents } from 'services/logrocket';
 import { entityCreateStoreSelectors } from 'stores/Create';
-import { CreateState, FormStatus } from 'stores/MiniCreate';
+import { EntityFormState, FormStatus } from 'stores/FormState';
 
 interface Props {
     closeLogs: Function;
@@ -20,6 +21,7 @@ interface Props {
     disabled: boolean;
     logEvent: CustomEvents.CAPTURE_TEST | CustomEvents.MATERIALIZATION_TEST;
     draftEditorStoreName: DraftEditorStoreNames;
+    formStateStoreName: FormStateStoreNames;
     detailsFormStoreName: DetailsFormStoreNames;
 }
 
@@ -29,32 +31,35 @@ function EntityTestButton({
     disabled,
     logEvent,
     draftEditorStoreName,
+    formStateStoreName,
     detailsFormStoreName,
 }: Props) {
-    const useEntityCreateStore = useRouteStore();
-    const showLogs = useZustandStore<
-        CreateState,
-        CreateState['formState']['showLogs']
-    >(detailsFormStoreName, (state) => state.formState.showLogs);
-
-    const logToken = useZustandStore<
-        CreateState,
-        CreateState['formState']['logToken']
-    >(detailsFormStoreName, (state) => state.formState.logToken);
-
-    const formStatus = useZustandStore<
-        CreateState,
-        CreateState['formState']['status']
-    >(detailsFormStoreName, (state) => state.formState.status);
-
-    const messagePrefix = useEntityCreateStore(
-        entityCreateStoreSelectors.messagePrefix
-    );
-
+    // Draft Editor Store
     const draftId = useZustandStore<
         EditorStoreState<DraftSpecQuery>,
         EditorStoreState<DraftSpecQuery>['id']
     >(draftEditorStoreName, (state) => state.id);
+
+    // Form State Store
+    const showLogs = useZustandStore<
+        EntityFormState,
+        EntityFormState['formState']['showLogs']
+    >(formStateStoreName, (state) => state.formState.showLogs);
+
+    const logToken = useZustandStore<
+        EntityFormState,
+        EntityFormState['formState']['logToken']
+    >(formStateStoreName, (state) => state.formState.logToken);
+
+    const formStatus = useZustandStore<
+        EntityFormState,
+        EntityFormState['formState']['status']
+    >(formStateStoreName, (state) => state.formState.status);
+
+    const useEntityCreateStore = useRouteStore();
+    const messagePrefix = useEntityCreateStore(
+        entityCreateStoreSelectors.messagePrefix
+    );
 
     return (
         <>
@@ -74,7 +79,7 @@ function EntityTestButton({
                 actionComponent={
                     <LogDialogActions
                         close={closeLogs}
-                        detailsFormStoreName={detailsFormStoreName}
+                        formStateStoreName={formStateStoreName}
                     />
                 }
             />
@@ -84,6 +89,7 @@ function EntityTestButton({
                 onFailure={callFailed}
                 logEvent={logEvent}
                 draftEditorStoreName={draftEditorStoreName}
+                formStateStoreName={formStateStoreName}
                 detailsFormStoreName={detailsFormStoreName}
             />
         </>
